refactor: extract build utils
